requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.14.1
mongoengine==0.24.0
packaging==21.3
pymongo==4.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
urllib3==1.26.8
webargs==8.1.0
Werkzeug==2.0.3
WTForms==3.0.1
